Initial README.md
Establishing the foundation for Mridun Gupta Codes with a comprehensive README.md outlining the mission, repository types, and guidelines for users and contributors. This commit includes details on both public and private repositories, as well as information on how users and contributors can get started with our code solutions. Contact details are also provided for inquiries, collaborations, and support.

Signed-off-by: Mridun Gupta <gmridun@gmail.com>